feat(plan399 Phase11.1): a2r int 类型 i32→i64 根治 + 黄金同步
纠正前期误判: a2r 实际有 327 单元测试 + 239 黄金用例(需 --features
test-trans), 非测试薄弱.

根因: trans/rust.rs:1162 Type::Int => "i32", 但后端 types.rs 用 i64
(auto_type_to_rust int→i64). 不一致靠 post_process_db_rs 的
code.replace("i32","i64") 暴力修补.

根治:
- rust_type_name Type::Int => "i64" (类型映射点)
- 批量 as i32 → as i64 (cast 语义, 写进输出的字面量)
- parse::<i32> → parse::<i64>
- 111 个 a2r + 42 个 cookbook 黄金文件(.expected.rs)同步更新(i32→i64)

验证: a2r 327 passed/0 failed (--features test-trans, 排除 concurrency
栈溢出). 黄金文件用 a2r 实际输出(.wrong.rs)作新基线确保准确.

待后续(需磁盘空间): 移除 post_process_db_rs 的 i32→i64 后处理 +
015/017 端到端编译验证.

+++ crates/auto-lang/test/a2r/04_strings/006_multi_fstr/multi_fstr.expected.rs
@@ -2,7 +2,7 @@
 
 fn main() {
     let name: String = "World".to_string();;
-    let age: i32 = 25;;
+    let age: i64 = 25;;
     let msg: String = format!("Hello {},
 you are {} years old.", name, age);;
     println!("{}", msg);
