+++ crates/auto-lang/test/a2r/09_option_result/001_option/option.expected.rs
@@ -2,7 +2,7 @@
 
 fn main() {
 
-    let x: i32 = 10;
+    let x: i64 = 10;
 
     match x {
         0 => println!("X is ZERO"),
