+++ crates/auto-lang/test/a2r/10_collections/002_list_storage/list_storage.expected.rs
@@ -15,7 +15,7 @@ struct Heap<T> {
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 struct InlineInt64 {
-    pub buffer: Vec<i32>,
+    pub buffer: Vec<i64>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -27,12 +27,12 @@ struct List<T, S> {
 fn main() {
 
     let heap_list = List<int, Heap>.new();
-    let heap_len = (heap_list.len() as i32);
+    let heap_len = (heap_list.len() as i64);
     let heap_cap = heap_list.capacity();
 
 
     let inline_list = List<int, InlineInt64>.new();
-    let inline_len = (inline_list.len() as i32);
+    let inline_len = (inline_list.len() as i64);
     let inline_cap = inline_list.capacity();
 
     return 0;
