+++ crates/auto-lang/test/a2r/15_type_conversion/004_box_arc/box_arc.expected.rs
@@ -2,7 +2,7 @@
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 struct Node {
-    pub value: i32,
+    pub value: i64,
 }
 
 fn main() {
