+++ crates/auto-lang/test/a2r/19_ownership/009_move_closure/move_closure.expected.rs
@@ -2,10 +2,10 @@
 
 fn main() {
     let name: String = "cosmic".to_string();
-    let count: i32 = 3;
+    let count: i64 = 3;
 
 
-    let f = move |n: i32| { println!("{}", n); };
+    let f = move |n: i64| { println!("{}", n); };
     f(count);
