+++ crates/auto-lang/test/cookbook/encoding/003_csv_read/csv_read.expected.rs
@@ -4,7 +4,7 @@ fn main() {
     let csv_str: String = "name,age\nAlice,30\nBob,25".to_string();
 
     let mut lines = csv_str.split("\n").collect::<Vec<_>>();
-    assert!((lines.len() as i32) == 3);
+    assert!((lines.len() as i64) == 3);
     println!("Name: Alice, Age: 30");
     println!("Name: Bob, Age: 25");
 }
