+++ crates/auto-lang/test/cookbook/text/007_from_str/from_str.expected.rs
@@ -2,7 +2,7 @@
 
 fn parse_parts(s: &str) {
     let mut parts = s.split(",").collect::<Vec<_>>();
-    assert!((parts.len() as i32) == 2);
+    assert!((parts.len() as i64) == 2);
 }
 
 fn main() {
